requirements.txt
litellm
python-dotenv
streamlit
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
sentence-transformers
chromadb
pypdf
pdf2image
pytesseract
Pillow